some unit tests

+++ backend/src/main/java/songbook/models/ReferenceRetained.java
@@ -5,7 +5,7 @@ import songbook.collections.models.ReferenceVolume;
 public class ReferenceRetained {
     public String title;
     public ReferenceVolume volume;
-    public short page;
+    public int page;
     public String author;
     public Key key;
 }
